+++ src/fields/ContentBlock.php
@@ -784,7 +784,13 @@ JS, [
      */
     public function afterElementPropagate(ElementInterface $element, bool $isNew): void
     {
-        $this->contentBlockManager()->maintainNestedElements($element, $isNew);
+        // Only do anything if we're working with a saved content block
+        /** @var ContentBlockElement $value */
+        $value = $element->getFieldValue($this->handle);
+        if ($value->id) {
+            $this->contentBlockManager()->maintainNestedElements($element, $isNew);
+        }
+
         parent::afterElementPropagate($element, $isNew);
     }